requirements.txt
streamlit 
python-dotenv
google-generativeai
langchain
PYPDF2
chromadb
langchain-groq
faiss-cpu
groq
langchain_google_genai
pypdf